backend/requirements.txt
fastapi
uvicorn

python-dotenv
pydantic-settings

langgraph
langgraph-checkpoint-sqlite
langchain-openai
langchain-tavily

SQLAlchemy
aiosqlite